refactor(app): fold app/src/utils into lib, split by concern
Two directories held the same kind of shared code: lib/ (many small,
single-concern files) and utils/ (a two-file grab-bag). Move utils/
into lib/, one file per concern - format-time.ts (also takes the
formatRelativeTime that lived beside the newly-centralized formatTime),
format-number.ts, and method-display.ts - and drop the deprecated
loadTestTypeToLabel shim next to the loadTestModeLabel it aliases in
constants/load-test-modes.ts, since constants/ already holds functions
there. Update every import site and drop the @/utils alias from
tsconfig.json and vite.config.ts.
