+++ app/src/components/layout/context-bar/RecentSendsSection.tsx
@@ -30,7 +30,7 @@
 import { useRecentDesignRunsQuery } from "@/queries";
 import { useTabsStore } from "@/stores";
 import { StatusCodeBadge, formatResponseTime } from "@/components/shared";
-import { formatRelativeTime } from "@/utils";
+import { formatRelativeTime } from "@/lib/format-time";
 import { SectionEmpty, SectionLoading } from "./Section";
 import type { ContextBarSectionProps } from "./types";
 import type { Run } from "@/types";
