+++ app/src/lib/format-number.ts
@@ -1,0 +1,22 @@
+/**
+ * Copyright (c) 2026 Atharva Kusumbia
+ *
+ * This source code is licensed under the Apache 2.0 license found in the
+ * LICENSE file in the "app" directory of this source tree.
+ */
+
+/**
+ * Format number with thousand separators
+ */
+export function formatNumber(num: number | undefined | null): string {
+	if (num === undefined || num === null || isNaN(num)) return "0";
+	return num.toLocaleString();
+}
+
+/**
+ * Format percentage
+ */
+export function formatPercent(value: number, total: number): string {
+	if (total === 0) return "0%";
+	return `${((value / total) * 100).toFixed(1)}%`;
+}
