+++ app/src/lib/format-time.ts
@@ -9,3 +9,24 @@
 export function formatTime(ms: number): string {
 	return new Date(ms).toLocaleTimeString();
 }
+
+/**
+ * Format date to relative time (e.g., "2 hours ago")
+ */
+export function formatRelativeTime(timestamp: string | number): string {
+	const date = new Date(timestamp);
+	const now = new Date();
+	const diffMs = now.getTime() - date.getTime();
+
+	const seconds = Math.floor(diffMs / 1000);
+	const minutes = Math.floor(seconds / 60);
+	const hours = Math.floor(minutes / 60);
+	const days = Math.floor(hours / 24);
+
+	if (seconds < 60) return "just now";
+	if (minutes < 60) return `${minutes}m ago`;
+	if (hours < 24) return `${hours}h ago`;
+	if (days < 7) return `${days}d ago`;
+
+	return date.toLocaleString();
+}
