+++ app/src/modules/dashboard/components/MetricsView.tsx
@@ -23,7 +23,7 @@
 
 import { memo, useMemo, useState } from "react";
 import { Activity } from "lucide-react";
-import { formatNumber } from "@/utils";
+import { formatNumber } from "@/lib/format-number";
 import { useDashboardStore } from "@/stores";
 import { useHostSleeps } from "@/stores/host-sleep-store";
 import type { MetricsViewProps, DashboardDerived } from "../types";
