+++ app/src/modules/dashboard/components/hero/DroppedRequestsCard.tsx
@@ -5,7 +5,7 @@
  * LICENSE file in the "app" directory of this source tree.
  */
 
-import { formatNumber } from "@/utils";
+import { formatNumber } from "@/lib/format-number";
 import { HeroCardShell, HeroValue } from "./HeroCardShell";
 
 /**
