+++ app/src/modules/history/main/components/OverviewTab.tsx
@@ -14,7 +14,7 @@
 import { AlertCircle } from "lucide-react";
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui";
 import { cn } from "@/lib/utils";
-import { formatNumber } from "@/utils";
+import { formatNumber } from "@/lib/format-number";
 import {
 	CapacitySummary,
 	ContractCoverage,
