--- app/src/utils/index.ts
@@ -1,8 +1,0 @@
-/**
- * Copyright (c) 2026 Atharva Kusumbia
- *
- * This source code is licensed under the Apache 2.0 license found in the
- * LICENSE file in the "app" directory of this source tree.
- */
-
-export * from "./helpers";
