+++ app/tsconfig.json
@@ -63,7 +63,6 @@
 			"@/hooks/*": ["./src/hooks/*"],
 			"@/services/*": ["./src/services/*"],
 			"@/types/*": ["./src/types/*"],
-			"@/utils/*": ["./src/utils/*"],
 			"@shared/*": ["../shared/*"]
 		}
 	},
