+++ app/vite.config.ts
@@ -38,7 +38,6 @@ export default defineConfig({
 			"@/hooks": path.resolve(__dirname, "./src/hooks"),
 			"@/services": path.resolve(__dirname, "./src/services"),
 			"@/types": path.resolve(__dirname, "./src/types"),
-			"@/utils": path.resolve(__dirname, "./src/utils"),
 			// Repo-level shared assets (icons). The renderer needs its own bundled
 			// copy - the electron-builder / build.py icons are filesystem artifacts
 			// the sandboxed web content cannot reach - but it reads them from the
